requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-contrib-postgres==0.0.1
django-cors-headers==4.0.0
django-filter==23.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-tree-queries==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.24.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
